feat: 로그인 여부에 따라 헤더 아이콘 표시 변경
